requirements.txt
accelerate==0.34.2
altair==5.4.1
attrs==24.2.0
beartype==0.18.5
bitsandbytes==0.42.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
einops==0.8.0
einx==0.3.0
ema-pytorch==0.6.2
filelock==3.16.0
frozendict==2.4.4
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
huggingface-hub==0.24.7
idna==3.9
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
memory-efficient-attention-pytorch==0.1.6
mpmath==1.3.0
muse-maskgit-pytorch==0.3.5
narwhals==1.8.0
networkx==3.3
numpy==2.1.1
packaging==24.1
pandas==2.2.2
pillow==10.4.0
protobuf==5.28.1
psutil==6.0.0
pyarrow==17.0.0
pydeck==0.9.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rich==13.8.1
rpds-py==0.20.0
safetensors==0.4.5
scipy==1.14.1
sentencepiece==0.2.0
six==1.16.0
smmap==5.0.1
streamlit==1.38.0
sympy==1.13.2
tenacity==8.5.0
tokenizers==0.19.1
toml==0.10.2
torch==2.4.1
torchvision==0.19.1
tornado==6.4.1
tqdm==4.66.5
transformers==4.44.2
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
vector-quantize-pytorch==1.17.3
